Separér Button og IconButton til to forskjellige stories
